+++ 10_storage/g_radm_2014.f90
@@ -30,7 +30,7 @@ module vars
     integer :: nh
     integer :: nx,ny ! grid dimensions
     integer :: ncty  ! number of point stations
-    integer :: idcf  ! ID cell in file
+    integer*8 :: idcf  ! ID cell in file
     integer :: zlev       ! Layer of emission (1 to 8) 8 lower 1 upper
     integer,parameter :: ipm=29  ! Posicion del archivo PM2.5
     integer,parameter :: icn=36    ! Posicion archivo CN del INEM
