+++ 10_storage/g_saprc_2014.f90
@@ -31,7 +31,7 @@ module varss
     integer :: nh    !  hours in a day
     integer :: nx,ny ! grid dimensions
     integer :: ncty  ! number of point stations
-    integer :: idcf  ! ID cell in file
+    integer*8 :: idcf  ! ID cell in file
     integer :: zlev       ! Layer of emission (1 to 8) 8 lower 1 upper
     integer,parameter :: ipm=40  ! Posicion del archivo PM2.5
     integer,parameter :: icn=47    ! Posicion archivo CN del INEM
